requirements.txt
# Application Dependencies
streamlit>=1.47.0
bcrypt>=4.3.0
python-dotenv>=1.1.0
cryptography>=45.0.0
plotly>=6.2.0
pandas>=2.2.0

# Testing Dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-html>=3.1.0      # HTML test reports
pytest-xdist>=3.0.0     # Parallel test execution
pytest-benchmark>=4.0.0  # Performance testing
freezegun>=1.2.0         # Time mocking for tests
faker>=15.0.0            # Generate fake test data

# Data Quality
great-expectations>=0.18.0
